requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
gradio
httpx
python-dotenv==1.0.0
langchain-community==0.0.38
langchain-text-splitters==0.0.1
langchain-openai==0.0.8
langchain-core==0.1.52
langchain==0.1.16
pypdf
chromadb